[S13P21A302-111] WebSocket 기반 실시간 채팅 구현

See merge request s13-bigdata-dist-sub1/S13P21A302!14

+++ backend-spring/leaper/src/main/java/com/ssafy/leaper/domain/chat/dto/request/ChatMessageSendRequest.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class ChatMessageSendRequest {
     @NotNull(message = "발신자 ID는 필수입니다.")
-    private Long senderId;
+    private Integer senderId;
     
     @NotBlank(message = "메시지는 필수입니다.")
     private String content;
